# webapp/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VooZuu App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/contacts/saveContact:
   # binds a127 app logic to a route
   x-swagger-router-controller: contact_ctrl
   post:
      tags: ['Contact Module']
      description: Save Contact
      # used as the method name of the controller
      operationId: saveContact
      parameters:
        - in: body
          name: body
          description: body
          required: true
          schema:
             $ref: '#/definitions/Contacts'          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/contacts/getContacts:
    x-swagger-router-controller: contact_ctrl
    get:
      tags: ['Contact Module']
      description: to fetch contact data
      # used as the method name of the controller
      operationId: getContacts
      parameters:
        - name: userId
          in: query
          description: user id
          required: true
          type: string        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # save tag comes here
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Contacts:
    type: array
    items: 
          $ref: '#/definitions/saveContact'
  saveContact:
    type: object
    discriminator: type  
    properties:
      contactId:
       type: string
      userId:
       type: string
      fullName:
       type: string
      familyName:
        type: string
      jobTitle:
       type: string
      company:
        type: string
      phoneNumbers:
        type: array
        items: 
          $ref: '#/definitions/phoneNumbers'
      emails:
        type: array
        items: 
          $ref: '#/definitions/emails'
...
  phoneNumbers:
    type: object
    discriminator: type
    properties:
      type:
        type: string
      number:
        type: string
    required:
      - type
      - number
  emails:
    type: object
    discriminator: type
    properties:
      type:
        type: string
      email:
        type: string
    required:
      - type
      - email